requirements.txt
gitdb~=4.0.11
Pillow~=9.5.0
pip~=21.3.1
toml~=0.10.2
wheel~=0.37.1
tornado~=6.2
rsa~=4.9
pyasn1~=0.5.1
toolz~=0.12.1
gast~=0.4.0
numpy~=1.21.6
h5py~=3.8.0
pytz~=2024.1
rich~=13.7.1
attrs~=23.2.0
pygments~=2.17.2
smmap~=5.0.1
pywin32~=306
keras~=2.11.0
tensorflow~=2.11.0
six~=1.16.0
scipy~=1.7.3
pandas~=1.3.5
wrapt~=1.16.0
requests~=2.31.0
mdurl~=0.1.2
setuptools~=60.2.0
matplotlib~=3.5.3
jsonschema~=4.17.3
altair~=5.0.1
jinja2~=3.1.4
cachetools~=5.3.3
certifi~=2024.6.2
urllib3~=2.0.7
packaging~=23.2
MarkupSafe~=2.1.5
joblib~=1.3.2
python-dateutil~=2.9.0.post0
pyarrow~=12.0.1
pydeck~=0.8.1b1
blinker~=1.6.3
tzdata~=2024.1
Pympler~=1.1
seaborn~=0.12.2
cycler~=0.11.0
scikit-learn~=1.0.2
threadpoolctl~=3.1.0
colorama~=0.4.6
tzlocal~=4.3.1
Markdown~=3.4.4
oauthlib~=3.2.2
idna~=3.7
tenacity~=8.2.3
watchdog~=3.0.0
Werkzeug~=2.2.3
fonttools~=4.38.0
pyparsing~=3.1.2
click~=8.1.7
streamlit~=1.23.1
termcolor~=2.3.0
pyrsistent~=0.19.3
kiwisolver~=1.4.5
tensorboard~=2.11.2
flatbuffers~=24.3.25
astunparse~=1.6.3
validators~=0.20.0
decorator~=5.1.1
platformdirs~=4.0.0
zipp~=3.15.0